fix download URLs
Some of the urls stated in the enclosure element of the RSS feed didn't work and resulted in a 502.
Now the attribute node is parsed and the url in the included json string is used to download the media.
This not only makes it tremendously faster, but I also didn't encounter a single 502 since the change.