add filename to archive db
This change was needed to enable sorted hosting of the files (feature to come in the future).
I've decided to not implement a migration path for existing archives. So to enable it fully,
one needs to recreate the archive, possibly missing some items that have gone off the feed
in the meantime. Sorry. This will most probably happen again in upcoming releases.